Move everything to template headers
