+++ src/app/clusters/basic-information/BasicInformationClusterImpl.h
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2026 Project CHIP Authors
+ *    Copyright (c) 2025-2026 Project CHIP Authors
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,18 +14,17 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
-#include <algorithm>
+#pragma once
 
 #include <app/AttributeValueDecoder.h>
 #include <app/AttributeValueEncoder.h>
-#include <app/SpecificationDefinedRevisions.h>
 #include <app/persistence/AttributePersistence.h>
 #include <app/persistence/String.h>
 #include <app/server-cluster/AttributeListBuilder.h>
 #include <app/server-cluster/DefaultServerCluster.h>
 #include <app/server-cluster/OptionalAttributeSet.h>
 #include <app/server-cluster/ServerClusterContext.h>
+#include <app/SpecificationDefinedRevisions.h>
 #include <clusters/BasicInformation/Attributes.h>
 #include <clusters/BasicInformation/ClusterId.h>
 #include <clusters/BasicInformation/Enums.h>
@@ -38,13 +37,94 @@
 #include <platform/CHIPDeviceError.h>
 #include <protocols/interaction_model/StatusCode.h>
 #include <tracing/macros.h>
-
-// not really required, however helps IDEs
-#include <app/clusters/basic-information/BasicInformationCluster.h>
+#include <algorithm>
 
 namespace chip {
 namespace app {
 namespace Clusters {
+
+/// This class provides a code-driven implementation for the Basic Information cluster,
+/// centralizing its logic and state.
+///
+/// It uses a Policy-based design to decouple from DeviceLayer and generic platform logic.
+template <typename Policy>
+class BasicInformationClusterImpl : public DefaultServerCluster, public Policy::DelegateBase
+{
+public:
+    using Context = typename Policy::Context;
+
+    using OptionalAttributesSet = chip::app::OptionalAttributeSet< //
+        BasicInformation::Attributes::ManufacturingDate::Id,       //
+        BasicInformation::Attributes::PartNumber::Id,              //
+        BasicInformation::Attributes::ProductURL::Id,              //
+        BasicInformation::Attributes::ProductLabel::Id,            //
+        BasicInformation::Attributes::SerialNumber::Id,            //
+        BasicInformation::Attributes::LocalConfigDisabled::Id,     //
+        BasicInformation::Attributes::Reachable::Id,               //
+        BasicInformation::Attributes::ProductAppearance::Id,       //
+        // Old specification versions had UniqueID as optional, so this
+        // appears here even though MANDATORY in the latest spec. We
+        // default it enabled (to decrease chances of error)
+        BasicInformation::Attributes::UniqueID::Id //
+        >;
+
+    BasicInformationClusterImpl(OptionalAttributesSet optionalAttributeSet, Context ctx) :
+        DefaultServerCluster({ kRootEndpointId, BasicInformation::Id }), mEnabledOptionalAttributes(optionalAttributeSet),
+        mPolicy(ctx)
+    {
+        mEnabledOptionalAttributes
+            .template Set<BasicInformation::Attributes::UniqueID::Id>(); // Unless told otherwise, unique id is mandatory
+    }
+
+    OptionalAttributesSet & OptionalAttributes() { return mEnabledOptionalAttributes; }
+
+    // Server cluster implementation
+    CHIP_ERROR Startup(ServerClusterContext & context) override;
+    void Shutdown(ClusterShutdownType type) override;
+    DataModel::ActionReturnStatus ReadAttribute(const DataModel::ReadAttributeRequest & request,
+                                                AttributeValueEncoder & encoder) override;
+    DataModel::ActionReturnStatus WriteAttribute(const DataModel::WriteAttributeRequest & request,
+                                                 AttributeValueDecoder & decoder) override;
+    CHIP_ERROR Attributes(const ConcreteClusterPath & path, ReadOnlyBufferBuilder<DataModel::AttributeEntry> & builder) override;
+
+    // Policy::DelegateBase implementation (PlatformManagerDelegate for DeviceLayer)
+    /**
+     * @brief Initialize the cluster
+     *
+     * This method attempts to register the cluster as the DeviceLayer::PlatformManagerDelegate
+     * to receive system shutdown events (OnShutDown).
+     * * NOTE: Registration is conditional. It will ONLY register this cluster as the delegate
+     * if the PlatformManager does not currently have a delegate set. If the application
+     * has already registered a delegate, this cluster will respect that configuration
+     * and will NOT overwrite it.
+     */
+    void OnStartUp(uint32_t softwareVersion) override;
+
+    void OnShutDown() override;
+
+    // ConfigurationVersionDelegate, however NOT overridable to save
+    // some flash in case this feature is never used. This means applications that may
+    // change configurations at runtime pay a bit more flash, however those are probably more
+    // dynamic (i.e. larger) systems like bridges or more complex systems.
+    CHIP_ERROR IncreaseConfigurationVersion();
+
+private:
+    // Implementation of just the write, without doing a attribute change notification.
+    //
+    // Will return ActionReturnStatus::FixedStatus::kWriteSuccessNoop if the attribute write
+    // is not changing the underlying data.
+    DataModel::ActionReturnStatus WriteImpl(const DataModel::WriteAttributeRequest & request, AttributeValueDecoder & decoder);
+
+    // Reads a single device info string and handles error sanitization (clearing the buffer)
+    // if the attribute is optional and not implemented.
+    template <typename EncodeFunction>
+    CHIP_ERROR ReadConfigurationString(EncodeFunction && getter, AttributeValueEncoder & encoder);
+
+    OptionalAttributesSet mEnabledOptionalAttributes;
+
+    Storage::String<32> mNodeLabel;
+    Policy mPolicy;
+};
 
 namespace BasicInformationClusterImplDetails {
 // Unique ID became mandatory in 4. If we have no unique id, claim revision 3
@@ -53,7 +133,7 @@ inline constexpr uint32_t kRevisionWithoutUniqueId = 3;
 // This is generally DeviceLayer::ConfigurationManager::kMaxLocationLength
 // However we try to not tie these constants to device layer.
 //
-// Technically we should static_assert(kExpectedFixedLocationLength == DeviceLayer::ConfigurationManager::kMaxLocationLength);
+// Technically we should static_assert(kExpectedFixedLocationLength == DeviceLayer::ConfigurationManager::kLocationLength);
 constexpr size_t kExpectedFixedLocationLength = 2;
 
 // The largest buffer size needed to read string attributes. This allows re-use of a common
@@ -69,7 +149,7 @@ constexpr size_t kMaxStringLength = std::max({
     BasicInformation::Attributes::SerialNumber::TypeInfo::MaxLength(),
 });
 
-inline CHIP_ERROR EncodeStringOnSuccess(CHIP_ERROR status, AttributeValueEncoder & encoder, const char * buf, size_t maxBufSize)
+    inline CHIP_ERROR EncodeStringOnSuccess(CHIP_ERROR status, AttributeValueEncoder & encoder, const char * buf, size_t maxBufSize)
 {
     ReturnErrorOnFailure(status);
     return encoder.Encode(chip::CharSpan(buf, strnlen(buf, maxBufSize)));
@@ -150,11 +230,6 @@ template <typename EncodeFunction>
 CHIP_ERROR BasicInformationClusterImpl<Policy>::ReadConfigurationString(EncodeFunction && getter, AttributeValueEncoder & encoder)
 {
     char buffer[BasicInformationClusterImplDetails::kMaxStringLength + 1];
-
-    // In case buggy getter returns success but does not change the string
-    // NOTE: we do this instead of `= {0}` to try to use less flash than a memset.
-    buffer[0] = 0;
-
     CHIP_ERROR status = getter(buffer, sizeof(buffer));
     return BasicInformationClusterImplDetails::EncodeStringOnSuccess(status, encoder, buffer,
                                                                      BasicInformationClusterImplDetails::kMaxStringLength);
@@ -226,15 +301,14 @@ DataModel::ActionReturnStatus BasicInformationClusterImpl<Policy>::ReadAttribute
         return ReadConfigurationString([this](char * buf, size_t size) { return mPolicy.GetSoftwareVersionString(buf, size); },
                                        encoder);
     case ManufacturingDate::Id: {
-        constexpr size_t kMaxLength                  = BasicInformation::Attributes::ManufacturingDate::TypeInfo::MaxLength();
-        constexpr size_t kMaxDateLength              = 8; // YYYYMMDD
-        char manufacturingDateString[kMaxLength + 1] = { 0 };
-
+        constexpr size_t kMaxDateLength  = 8;     // YYYYMMDD
+        char manufacturingDateString[17] = { 0 }; // kMaxManufacturingDateLength around 16
         uint16_t manufacturingYear;
         uint8_t manufacturingMonth;
         uint8_t manufacturingDayOfMonth;
         size_t totalManufacturingDateLen = 0;
-        MutableCharSpan vendorSuffixSpan(manufacturingDateString + kMaxDateLength, kMaxLength - kMaxDateLength);
+        MutableCharSpan vendorSuffixSpan(manufacturingDateString + kMaxDateLength,
+                                         sizeof(manufacturingDateString) - kMaxDateLength);
         CHIP_ERROR status = mPolicy.GetManufacturingDate(manufacturingYear, manufacturingMonth, manufacturingDayOfMonth);
 
         if (status == CHIP_DEVICE_ERROR_CONFIG_NOT_FOUND || status == CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE)
